Added a products activity

+++ app/src/main/java/com/somnium/simplyshop/entities/Product.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 class Product(
         var name: String,
         var description: String,
-        var price: Int
+        var price: Float
 ): Serializable
